templates/diffusers/requirements.txt
fastapi>=0.104.0
pydantic>=2.0.0
torch>=2.0.0
diffusers>=0.24.0
#transformers>=4.35.0
accelerate>=0.24.0
#ray[serve]>=2.7.0
Pillow>=10.0.0
requests>=2.31.0
uvicorn>=0.24.0